requirements.txt
# =============================================================================
# DeepEval Examples - Python Dependencies
# =============================================================================

# -----------------------------------------------------------------------------
# LangChain Ecosystem
# -----------------------------------------------------------------------------
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# -----------------------------------------------------------------------------
# LangGraph (for Agent example)
# -----------------------------------------------------------------------------
langgraph>=0.2.0

# -----------------------------------------------------------------------------
# Vector Store
# -----------------------------------------------------------------------------
faiss-cpu>=1.7.0

# -----------------------------------------------------------------------------
# Evaluation
# -----------------------------------------------------------------------------
deepeval>=2.0.0

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
python-dotenv>=1.0.0
